requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.11
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
Pillow==7.0.0
pkg-resources==0.0.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
